Bug 1362167 - Split gfxShapedText.mFlags into two 16-bit flags fields, and arrange storage more compactly to reduce size of gfxShapedWord and gfxTextRun objects. r=jrmuizel
UltraBlame original commit: 0cbf75382a9abb55be44e22cbc351e9214f4431f

+++ layout/base/nsLayoutUtils.h
@@ -1997,7 +1997,7 @@ public:
 
 
 
-  static uint32_t GetTextRunFlagsForStyle(nsStyleContext* aStyleContext,
+  static uint16_t GetTextRunFlagsForStyle(nsStyleContext* aStyleContext,
                                           const nsStyleFont* aStyleFont,
                                           const nsStyleText* aStyleText,
                                           nscoord aLetterSpacing);
@@ -2005,7 +2005,7 @@ public:
   
 
 
-  static uint32_t GetTextRunOrientFlagsForStyle(nsStyleContext* aStyleContext);
+  static uint16_t GetTextRunOrientFlagsForStyle(nsStyleContext* aStyleContext);
